dms/requirements.txt
Flask==3.0.0
pymongo==4.8.0
python-dotenv==1.0.0
APScheduler==3.10.4
geopy==2.4.1
numpy==1.26.4
flask-cors==4.0.0
